refactor: quickview as a language feature
