+++ src/brackets.js
@@ -172,6 +172,7 @@ define(function (require, exports, module) {
     //load language features
     require("features/ParameterHintsManager");
     require("features/JumpToDefManager");
+    require("features/QuickViewManager");
 
     // Load modules that self-register and just need to get included in the main project
     require("command/DefaultMenus");
