--- src/extensions/default/QuickView/QuickViewTemplate.html
@@ -1,4 +1,0 @@
-<div id="quick-view-container">
-    <div class="preview-content">
-    </div>
-</div>
